Renamed inputstream.js to inputStream.js | Code cleanup
