+++ cli.js
@@ -1,16 +1,16 @@
 #!/usr/bin/env node
 
-//since the interpreter works through this file(which is in the root directory), we can create a global variable for the app directory here
-global.appRoot = process.cwd();
+//since the interpreter works through this file, we can create a global variable for the app directory and the path module here
+global.path = require("path");
+global.rootDir = path.dirname(__filename);
 
 //we are telling *nix systems that the interpreter of our JavaScript file should be /usr/bin/env node which looks up for the locally-installed node executable.
 const packageJson = require("./package.json");
-const path = require("path");
-const InputStream = require("./inputstream.js");
-const Lexer = require("./lexer.js");
-const Parser = require("./parsers/parser.js");
-const Environment = require("./environment.js");
-const MainInterpreter = require("./interpreters/maininterpreter.js");
+const _inputStream = require("./inputStream.js");
+const _lexer = require("./lexer.js");
+const _parser = require("./parsers/parser.js");
+const _environment = require("./environment.js");
+const _mainInterpreter = require("./interpreters/maininterpreter.js");
 const constants = require("./constants.js");
 
 //process.argv will usually have length two, the zeroth item being the "node" interpreter 
@@ -21,8 +21,8 @@ const filename = process.argv[2];
 if (filename === "-v") {
     console.log(packageJson.version);
 } else if (path.extname(filename) === constants.YL_EXT) {
-    const parser = new Parser(new Lexer(new InputStream(filename)));
-    new MainInterpreter(new Environment(), parser).interpreteProgram();
+    const parser = new _parser(new _lexer(new _inputStream(filename)));
+    new _mainInterpreter(new _environment(), parser).interpreteProgram();
 } else {
     throw "Invalid Yorlang command line argument";
 }
