+++ inputStream.js
